`profanity` enhancements.
adds 'include' and 'exclude' to parsed rules, allowing rules to only apply to certain files, or not apply to certain files.

adds 'notContains' as a rule type, allowing rules to test that a file must contain a string.

adds base `include` and `exclude` cli flags, removes the `filter` cli flag.